Make layouts more flexible to prepare for hotel booking flow
