+++ src/components/funnels/funnels/register/layout/form/full-width.tsx
@@ -1,0 +1,20 @@
+import { ReactNode } from 'react'
+import FullWidthFunnelLayout from '~/components/funnels/layout/full-width'
+import RegisterHeader from '../header'
+
+export interface FullWidthRegisterFunnelLayoutProps {
+	children: ReactNode
+	currentStep: number
+	onNext: () => void
+}
+
+const FullWidthRegisterFunnelLayout = ({ children, currentStep, onNext }: FullWidthRegisterFunnelLayoutProps) =>
+	<FullWidthFunnelLayout
+		header={<RegisterHeader currentStep={currentStep}/>}
+		isFirstPage={currentStep === 0}
+		onNext={onNext}
+	>
+		{children}
+	</FullWidthFunnelLayout>
+
+export default FullWidthRegisterFunnelLayout
