+++ src/components/funnels/funnels/register/steps/ticket.tsx
