+++ src/components/funnels/funnels/register/steps/ticket/level/price.tsx
