+++ src/components/funnels/layout/common.tsx
@@ -1,0 +1,21 @@
+/** @jsxImportSource @emotion/react */
+
+import { ReactNode } from 'react'
+import Footer from './footer'
+
+export interface CommonFunnelLayoutProps {
+	readonly children: ReactNode
+	readonly header?: ReactNode
+	readonly isFirstPage?: boolean
+	readonly onNext: () => void
+}
+
+const CommonFunnelLayout = ({ children, header: headerContent, isFirstPage = false, onNext }: CommonFunnelLayoutProps) => <>
+	<header>
+		{headerContent}
+	</header>
+	{children}
+	<Footer canBack={!isFirstPage} onNext={onNext}/>
+</>
+
+export default CommonFunnelLayout
