+++ src/components/funnels/layout/footer.tsx
