+++ src/components/funnels/layout/full-width.tsx
@@ -1,0 +1,1 @@
+export { default, CommonFunnelLayoutProps as FullWidthFunnelLayoutProps } from './common'
