+++ src/navigation/router.tsx
@@ -2,13 +2,13 @@ import { Router } from '@reach/router'
 import NotFoundPage from '~/pages/404'
 import { IndexPage } from '~/pages/home'
 import RegisterPage from '~/pages/register'
-import Ticket from '~/components/register/steps/ticket'
-import TicketType from '~/components/register/steps/ticket/type'
-import TicketDay from '~/components/register/steps/ticket/day'
-import TicketLevel from '~/components/register/steps/ticket/level'
-import Personal from '~/components/register/steps/personal'
-import Contact from '~/components/register/steps/contact'
-import Optional from '~/components/register/steps/optional'
+import Ticket from '~/components/funnels/funnels/register/steps/ticket'
+import TicketType from '~/components/funnels/funnels/register/steps/ticket/type'
+import TicketDay from '~/components/funnels/funnels/register/steps/ticket/day'
+import TicketLevel from '~/components/funnels/funnels/register/steps/ticket/level'
+import Personal from '~/components/funnels/funnels/register/steps/personal'
+import Contact from '~/components/funnels/funnels/register/steps/contact'
+import Optional from '~/components/funnels/funnels/register/steps/optional'
 
 import * as ROUTES from './routes'
