+++ src/state/actions/register.ts
@@ -2,14 +2,14 @@ import type { DeepPartial } from 'ts-essentials'
 import { TicketType, TicketLevel, ContactInfo, OptionalInfo, PersonalInfo } from '~/state/models/register'
 import { createAction } from './create-action'
 
-export const ChangeTicketType = createAction<DeepPartial<TicketType['type']>, '[Register] Change ticket type'>('[Register] Change ticket type')
-export const ChangeTicketDay = createAction<DeepPartial<string>, '[Register] Change ticket day'>('[Register] Change ticket day')
+export const ChangeTicketType = createAction<DeepPartial<{ type: TicketType['type'] }>, '[Register] Change ticket type'>('[Register] Change ticket type')
+export const ChangeTicketDay = createAction<DeepPartial<{ day: string }>, '[Register] Change ticket day'>('[Register] Change ticket day')
 export const ChangeTicketLevel = createAction<DeepPartial<TicketLevel>, '[Register] Change ticket level'>('[Register] Change ticket level')
 export const ChangeContactInfo = createAction<DeepPartial<ContactInfo>, '[Register] Change contact info'>('[Register] Change contact info')
 export const ChangeOptionalInfo = createAction<DeepPartial<OptionalInfo>, '[Register] Change optional info'>('[Register] Change optional info')
 export const ChangePersonalInfo = createAction<DeepPartial<PersonalInfo>, '[Register] Change personal info'>('[Register] Change personal info')
-export const SubmitTicketType = createAction<TicketType['type'], '[Register] Submit ticket type'>('[Register] Submit ticket type')
-export const SubmitTicketDay = createAction<string, '[Register] Submit ticket day'>('[Register] Submit ticket day')
+export const SubmitTicketType = createAction<{ type: TicketType['type'] }, '[Register] Submit ticket type'>('[Register] Submit ticket type')
+export const SubmitTicketDay = createAction<{ day: string }, '[Register] Submit ticket day'>('[Register] Submit ticket day')
 export const SubmitTicketLevel = createAction<TicketLevel, '[Register] Submit ticket level'>('[Register] Submit ticket level')
 export const SubmitContactInfo = createAction<ContactInfo, '[Register] Submit contact info'>('[Register] Submit contact info')
 export const SubmitOptionalInfo = createAction<OptionalInfo, '[Register] Submit optional info'>('[Register] Submit optional info')
